:recycle: refactor: Interest DTO, VO 리팩토링
