+++ src/main/java/sandbox/apricot/interest/dto/request/InterestRegister.java
@@ -1,22 +1,23 @@
 package sandbox.apricot.interest.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
+import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
 @Builder
+@RequiredArgsConstructor
 public class InterestRegister {
 
-    @NotBlank(message = "분야(카테고리) 식별 정보가 필요 합니다.")
-    private final String categoryId; // 분야 고유 식별(FK)
+    private final Long memberId;
+    private final List<String> categoryInfo;
 
-    @NotBlank(message = "우선 순위 정보가 필요 합니다.")
-    private final int priority; // 우선 순위
-
-    public InterestRegister(String categoryId, int priority) {
-        this.categoryId = categoryId;
-        this.priority = priority;
+    public static InterestRegister of(Long memberId, List<String> categoryInfo) {
+        return InterestRegister.builder()
+                .memberId(memberId)
+                .categoryInfo(categoryInfo)
+                .build();
     }
 
 }
